--- src/analyzers/patchers/__init__.py
@@ -1,1 +1,0 @@
-ATTACH_DIR = "src/analyzers/patchers/attachments/"
